% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemanipulation.R
\name{phylo_drop_short_branches}
\alias{phylo_drop_short_branches}
\title{Function to drop very short terminal branches}
\usage{
phylo_drop_short_branches(phy, threshold = 0.01)
}
\arguments{
\item{phy}{A phylo object}

\item{threshold}{The maximum length of the branch to drop}
}
\description{
This will drop the shortest terminal branch, collapse any nodes that only have one descendant, and repeat. The goal is to get rid of terminal polytomies.
}
